Implemented database cleaning

+++ src/engine/frontend/win/editor/gridEditor/gridEditorDoc.cpp
@@ -129,7 +129,7 @@ bool CGridEditor::LoadDocument(const wxObjectDataPtr<CBackendSpreadsheetObject>&
 			if (row_size == nullptr)
 				continue;
 
-			if (row_size->m_row >= wxGridExt::GetNumberRows())
+			if ((int)row_size->m_row >= wxGridExt::GetNumberRows())
 				wxGridExt::AppendCols((int)row_size->m_row - wxGridExt::GetNumberRows() + 1);
 
 			wxGridExt::SetRowSize(row_size->m_row, row_size->m_height, 1.0f, false);
@@ -140,7 +140,7 @@ bool CGridEditor::LoadDocument(const wxObjectDataPtr<CBackendSpreadsheetObject>&
 			if (col_size == nullptr)
 				continue;
 
-			if (col_size->m_col >= wxGridExt::GetNumberCols())
+			if ((int)col_size->m_col >= wxGridExt::GetNumberCols())
 				wxGridExt::AppendCols((int)col_size->m_col - wxGridExt::GetNumberCols() + 1);
 
 			wxGridExt::SetColSize(col_size->m_col, col_size->m_width, 1.0f, false);
@@ -268,7 +268,7 @@ bool CGridEditor::LoadSpreadsheet(const CSpreadsheetDescription& spreadsheetDesc
 			if (row_size == nullptr)
 				continue;
 
-			if (row_size->m_row >= wxGridExt::GetNumberRows())
+			if ((int)row_size->m_row >= wxGridExt::GetNumberRows())
 				wxGridExt::AppendCols((int)row_size->m_row - wxGridExt::GetNumberRows() + 1);
 
 			wxGridExt::SetRowSize((int)row_size->m_row, row_size->m_height, 1.0f, false);
@@ -279,7 +279,7 @@ bool CGridEditor::LoadSpreadsheet(const CSpreadsheetDescription& spreadsheetDesc
 			if (col_size == nullptr)
 				continue;
 
-			if (col_size->m_col >= wxGridExt::GetNumberCols())
+			if ((int)col_size->m_col >= wxGridExt::GetNumberCols())
 				wxGridExt::AppendCols((int)col_size->m_col - wxGridExt::GetNumberCols() + 1);
 			
 			wxGridExt::SetColSize(col_size->m_col, col_size->m_width, 1.0f, false);
